requirements.txt
asgiref==3.9.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2025.8.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.2.1
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==46.0.5
daphne==4.1.2
Django==5.2.5
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
msgpack==1.1.2
mysqlclient==2.2.8
packaging==26.0
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyMySQL==1.1.2
pyngrok==7.3.0
pyOpenSSL==25.3.0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
redis==7.3.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wheel==0.45.1
zope.interface==8.2
